+++ src/instance.rs
@@ -1,9 +1,10 @@
 use lazy_static::lazy_static;
-use rutie::{class, methods, wrappable_struct, Hash, Module, Object, RString, Symbol};
+use rutie::{class, methods, wrappable_struct, AnyObject, Hash, Module, Object, RString};
 use std::collections::HashMap;
 use std::fs;
 use wasmtime as w;
 
+use crate::export::Export;
 use crate::func::Func;
 use crate::memory::Memory;
 
@@ -27,31 +28,25 @@ impl Instance {
         Instance { instance }
     }
 
-    fn exports(&self) -> (HashMap<String, Func>, HashMap<String, Memory>) {
-        let mut funcs = HashMap::new();
-        let mut memories = HashMap::new();
+    fn exports(&self) -> HashMap<String, Export> {
+        let mut exports = HashMap::new();
 
         for export in self.instance.exports() {
             match export.ty() {
                 w::ExternType::Func(_) => {
                     let name = export.name().to_string();
                     let func = Func::new(export.into_func().expect("failed to create func"));
-                    funcs.insert(name, func);
+                    exports.insert(name, Export::Func(func));
                 }
                 w::ExternType::Memory(_) => {
                     let memory = Memory::new();
-                    memories.insert(export.name().to_string(), memory);
+                    exports.insert(export.name().to_string(), Export::Memory(memory));
                 }
                 _ => {}
             }
         }
 
-        (funcs, memories)
-    }
-
-    pub fn funcs(&self) -> HashMap<String, Func> {
-        let (functions, _) = self.exports();
-        functions
+        exports
     }
 
     pub fn into_ruby(self) -> RubyInstance {
@@ -73,12 +68,14 @@ methods!(
         Instance::new(path.expect("failed read path").to_string()).into_ruby()
     }
 
-    fn ruby_instance_funcs() -> Hash {
-        let mut funcs = Hash::new();
-        for (export_name, func) in itself.get_data(&*INSTANCE_WRAPPER).funcs().into_iter() {
-            funcs.store(Symbol::new(&export_name), func.into_ruby());
+    fn ruby_instance_exports() -> Hash {
+        let mut exports = Hash::new();
+
+        for (export_name, export) in itself.get_data(&*INSTANCE_WRAPPER).exports().into_iter() {
+            exports.store::<RString, AnyObject>(RString::new_utf8(&export_name), export.into());
         }
-        funcs
+
+        exports
     }
 );
 
@@ -88,7 +85,7 @@ pub fn ruby_init() {
             .define_nested_class("Instance", None)
             .define(|class| {
                 class.def_self("new", ruby_instance_new);
-                class.def("funcs", ruby_instance_funcs);
+                class.def("exports", ruby_instance_exports);
             });
     });
 }